+++ admin/type-delete.php
@@ -1,16 +1,36 @@
-<?php require_once('top.php'); ?>
+<?php
+
+use Money\Exchange;
+
+ require_once('top.php'); ?>
 
 <?php
-    $id = $_GET['id'];
+   try {
+             $id = $_GET['id'];
 
-    $statement = $conn->prepare("DELETE FROM types WHERE id=?");
-    $statement->execute([$id]);
+            // Checking if there is an active property under location already. 
+            $statement = $conn->prepare("SELECT * FROM property WHERE type_id=?");
+            $statement->execute([$_GET['id']]);
+            $total = $statement->rowCount();
+            if($total > 0)
+            {
+                throw new Exception("There are some properties under this types. So you can not delete this type");
+            }
 
-    $success_message = 'Type is deleted successfully';
+            $statement = $conn->prepare("DELETE FROM types WHERE id=?");
+            $statement->execute([$id]);
 
-    $_SESSION['success_message'] = $success_message;
+            $success_message = 'Type is deleted successfully';
 
-    header('location: ' . ADMIN_URL . 'type-view.php');
-    exit
+            $_SESSION['success_message'] = $success_message;
+
+            header('location: ' . ADMIN_URL . 'type-view.php');
+            exit;
+   } catch (Exception $e) {
+        $error_message = $e->getMessage();
+        $_SESSION['error_message'] = $error_message;
+        header('location: ' . ADMIN_URL . 'location-view.php');
+        exit;
+   }
 ?>
